BDcontroler.py
import mysql.connector

class BDcontroler:
    def __init__(self):
        self.host = '35.239.140.3'
        self.user = 'root'
        self.password = 'root'
        self.database = 'music-system'
        self.connection = None
        self.cursor = None

    def connect(self):
        self.connection = mysql.connector.connect(
            host=self.host,
            user=self.user,
            password=self.password,
            database=self.database
        )
        self.cursor = self.connection.cursor()

    def execute_query(self, query):
        self.cursor.execute(query)
        return self.cursor.fetchall()

    def close(self):
        if self.cursor:
            self.cursor.close()
        if self.connection:
            self.connection.close()

