requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
requests==2.28.2
soupsieve==2.4.1
urllib3==1.26.15
Werkzeug==2.2.3
google-api-python-client==2.11.0
mysql-connector-python==8.0.26
flask_login==0.6.2